Update service configuration to be immutable
Currently the service configuration that is parsed from the JSON
configuration file for each service can be modified after its
creation. This means that after the configuration has been provided
to a module the configuration can still change, even if the module
itself makes no modifications directly. This can result in unexpected
behaviours.

This change makes the service configuration immutable. There is also
a smaller change to remove the HTTP module service configuration which
extends but makes no changes to the normal service configuration.