Add getting started video to auth SDK
Currently we do not link the video providing an end to end example
of using Keycloak in our documentation. This change adds the video
in the form of an image that once clicked will open YouTube.

The reason for not using the asciidoc video tag is because it does
not create the desired effect when rendered on GitHub.